requirements.txt
customtkinter>=5.2.0
numpy
openai==0.27.8
pynput==1.7.6
PyQt6>=6.4.0
python-dotenv==1.0.0
scipy>=1.9.0
sounddevice==0.4.6